wip: migrate session ending to new architecture
