+++ packages/graphql/src/graphql/ops/MEndLiveQuiz.graphql
@@ -1,0 +1,6 @@
+mutation EndLiveQuiz($id: String!) {
+  endLiveQuiz(id: $id) {
+    id
+    status
+  }
+}
