--- packages/graphql/src/graphql/ops/MEndSession.graphql
@@ -1,6 +1,0 @@
-mutation EndSession($id: String!) {
-  endSession(id: $id) {
-    id
-    status
-  }
-}
