+++ packages/graphql/src/public/client.json
@@ -40,8 +40,8 @@
    "EditTag": "92b05725c63972616bcf40f7c56b921af4308f6ef4583e90995a587c22d44bbd",
    "EnableCourseGamification": "c8a6b49c35ed2f0b3e2057204c4041a96689d4d333723a44bf7452350b11a6e3",
    "EndGroupActivity": "d70ab21634f5ec8a99e56fcbbcc3680b10c0c6dae53910b0277173c98d6a52e9",
+   "EndLiveQuiz": "f0cbd4410a7f9c868ce24562c06461349d490eaa787a41f0888f2e59246dbbd9",
    "EndMicroLearning": "a1ccbe55e2f4fdd33d0fb87f546c425770a6afe72451b7753079821a0f4a7310",
-   "EndSession": "c0d818931847f23c83c90e6fd5616b4724031fb69877de717106cc9a0a30ec46",
    "ExtendGroupActivity": "c3788f2608170940bedbec90fe77a623fb05249e20cd11ad2e4a529b5068149a",
    "ExtendMicroLearning": "bf40877047d36fc8549bf3050e75e858066f7a8515c97ad1207af5c3c3e227a6",
    "FinalRandomGroupAssignments": "eec2853f480a455126b4d00c0200b5f9a47a551d9d8b02b50806c31eb31cb8f7",
