+++ packages/graphql/src/public/schema.graphql
@@ -1057,8 +1057,8 @@ type Mutation {
   editTag(id: Int!, name: String!): Tag
   enableCourseGamification(courseId: String!): Course
   endGroupActivity(id: String!): GroupActivity
+  endLiveQuiz(id: String!): LiveQuiz
   endMicroLearning(id: String!): MicroLearning
-  endSession(id: String!): Session
   extendGroupActivity(endDate: Date!, id: String!): GroupActivity
   extendMicroLearning(endDate: Date!, id: String!): MicroLearning
   finalRandomGroupAssignments: Boolean!
