+++ packages/graphql/src/public/server.json
@@ -40,8 +40,8 @@
    "92b05725c63972616bcf40f7c56b921af4308f6ef4583e90995a587c22d44bbd": "mutation EditTag($id: Int!, $name: String!) {\n  editTag(id: $id, name: $name) {\n    id\n    name\n    __typename\n  }\n}",
    "c8a6b49c35ed2f0b3e2057204c4041a96689d4d333723a44bf7452350b11a6e3": "mutation EnableCourseGamification($courseId: String!) {\n  enableCourseGamification(courseId: $courseId) {\n    id\n    isGamificationEnabled\n    __typename\n  }\n}",
    "d70ab21634f5ec8a99e56fcbbcc3680b10c0c6dae53910b0277173c98d6a52e9": "mutation EndGroupActivity($id: String!) {\n  endGroupActivity(id: $id) {\n    id\n    status\n    scheduledEndAt\n    __typename\n  }\n}",
+   "f0cbd4410a7f9c868ce24562c06461349d490eaa787a41f0888f2e59246dbbd9": "mutation EndLiveQuiz($id: String!) {\n  endLiveQuiz(id: $id) {\n    id\n    status\n    __typename\n  }\n}",
    "a1ccbe55e2f4fdd33d0fb87f546c425770a6afe72451b7753079821a0f4a7310": "mutation EndMicroLearning($id: String!) {\n  endMicroLearning(id: $id) {\n    id\n    scheduledEndAt\n    __typename\n  }\n}",
-   "c0d818931847f23c83c90e6fd5616b4724031fb69877de717106cc9a0a30ec46": "mutation EndSession($id: String!) {\n  endSession(id: $id) {\n    id\n    status\n    __typename\n  }\n}",
    "c3788f2608170940bedbec90fe77a623fb05249e20cd11ad2e4a529b5068149a": "mutation ExtendGroupActivity($id: String!, $endDate: Date!) {\n  extendGroupActivity(id: $id, endDate: $endDate) {\n    id\n    scheduledEndAt\n    __typename\n  }\n}",
    "bf40877047d36fc8549bf3050e75e858066f7a8515c97ad1207af5c3c3e227a6": "mutation ExtendMicroLearning($id: String!, $endDate: Date!) {\n  extendMicroLearning(id: $id, endDate: $endDate) {\n    id\n    scheduledEndAt\n    __typename\n  }\n}",
    "eec2853f480a455126b4d00c0200b5f9a47a551d9d8b02b50806c31eb31cb8f7": "mutation FinalRandomGroupAssignments {\n  finalRandomGroupAssignments\n}",
