+++ packages/graphql/src/services/sessions.ts
@@ -7,341 +7,11 @@ import { max, mean, median, min, quantileSeq, std } from 'mathjs'
 import { createHmac } from 'node:crypto'
 import { mapValues, omitBy, prop, sortBy } from 'remeda'
 import type { Context, ContextWithUser } from '../lib/context.js'
-import { sendTeamsNotifications } from '../lib/util.js'
-
-// TODO: rework scheduling for serverless
-const scheduledJobs: Record<string, any> = {}
 
 // FIXME: move to config file or environment variable?
 const FIRST_ACHIEVEMENT_ID = 5
 const SECOND_ACHIEVEMENT_ID = 6
 const THIRD_ACHIEVEMENT_ID = 7
-
-interface EndSessionArgs {
-  id: string
-}
-
-export async function endSession({ id }: EndSessionArgs, ctx: ContextWithUser) {
-  const session = await ctx.prisma.liveSession.findFirst({
-    where: {
-      id,
-      ownerId: ctx.user.sub,
-    },
-    include: {
-      blocks: {
-        include: {
-          instances: {
-            orderBy: {
-              order: 'asc',
-            },
-          },
-        },
-        orderBy: {
-          id: 'asc',
-        },
-      },
-    },
-  })
-
-  // if there is no session matching the current user and session id, exit early
-  if (!session) {
-    return null
-  }
-
-  if (session.status === SessionStatus.COMPLETED) {
-    return session
-  }
-  if (
-    session.status === SessionStatus.PREPARED ||
-    session.status === SessionStatus.SCHEDULED
-  ) {
-    return null
-  }
-
-  try {
-    const sessionLB = await ctx.redisExec.hgetall(`s:${id}:lb`)
-    const sessionXP = await ctx.redisExec.hgetall(`s:${id}:xp`)
-
-    let promises: any[] = []
-
-    const participants: Record<string, any> = {}
-
-    Object.entries(sessionXP).forEach(([id, xp]) => {
-      participants[id] = {
-        xp: parseInt(xp),
-      }
-    })
-
-    Object.entries(sessionLB).forEach(([id, score]) => {
-      participants[id] = {
-        ...(participants[id] ?? {}),
-        score: parseInt(score),
-      }
-    })
-
-    // sessionXP should always be around as soon as there are logged-in participants (check first)
-    // sessionLB only for sessions that are compatible with points collection (check second)
-    if (sessionXP) {
-      let existingParticipants = (
-        await Promise.allSettled(
-          Object.entries(participants).map(async ([id, { score, xp }]) => {
-            const participant = await ctx.prisma.participant.findUnique({
-              where: { id },
-              include: {
-                // if the session is part of a course, include the corresponding participations
-                // if the participant is not part of the relevant course, the joined array will be empty
-                participations: session.courseId
-                  ? {
-                      where: {
-                        courseId: session.courseId,
-                      },
-                    }
-                  : undefined,
-              },
-            })
-
-            if (!participant) return null
-
-            return {
-              id,
-              score,
-              xp,
-              hasParticipation: participant.participations?.[0]?.isActive,
-            }
-          })
-        )
-      ).flatMap((result) => {
-        if (result.status !== 'fulfilled' || !result.value) return []
-        return [result.value]
-      })
-
-      // track the achievement ids, which should be awarded to the participants
-      let newAchievements: Record<string, number> = {}
-
-      // only award achievements, if the session did contain questions with sample solutions and at least three participants collected points
-      const awardAchievements = session.blocks.some(
-        (block) =>
-          block.instances.some((instance) => {
-            return instance.questionData.options.hasSampleSolution ?? false
-          }) &&
-          existingParticipants.filter(
-            ({ score }) => typeof score !== 'undefined'
-          ).length >= 3
-      )
-
-      // award achievements to the top 3 participants (and all others with equal scores)
-      if (awardAchievements) {
-        const topScores = existingParticipants
-          .filter(({ score }) => typeof score !== 'undefined')
-          .sort((a, b) => Number(b.score) - Number(a.score))
-          .slice(0, 3)
-
-        const firstRankAchievement = await ctx.prisma.achievement.findUnique({
-          where: { id: FIRST_ACHIEVEMENT_ID },
-        })
-        const secondRankAchievement = await ctx.prisma.achievement.findUnique({
-          where: { id: SECOND_ACHIEVEMENT_ID },
-        })
-        const thirdRankAchievement = await ctx.prisma.achievement.findUnique({
-          where: { id: THIRD_ACHIEVEMENT_ID },
-        })
-
-        const goldScore = topScores[0]?.score
-        const silverScore = topScores[1]?.score
-        const bronzeScore = topScores[2]?.score
-
-        // awarding logic (including point and xp updates):
-        // award gold to every participant with gold score
-        // award silver to every participant with silver score, if silver score != gold score
-        // award bronze to every participant with bronze score, if bronze score != silver score
-        existingParticipants = existingParticipants.map((participant) => {
-          if (
-            typeof participant.score === 'undefined' ||
-            typeof participant.xp === 'undefined'
-          ) {
-            return participant
-          }
-
-          if (participant.score === goldScore) {
-            participant.xp += firstRankAchievement!.rewardedXP ?? 0
-            participant.score += firstRankAchievement!.rewardedPoints ?? 0
-            newAchievements[participant.id] = firstRankAchievement!.id
-          }
-          if (participant.score === silverScore && silverScore !== goldScore) {
-            participant.xp += secondRankAchievement!.rewardedXP ?? 0
-            participant.score += secondRankAchievement!.rewardedPoints ?? 0
-            newAchievements[participant.id] = secondRankAchievement!.id
-          }
-          if (
-            participant.score === bronzeScore &&
-            bronzeScore !== silverScore
-          ) {
-            participant.xp += thirdRankAchievement!.rewardedXP ?? 0
-            participant.score += thirdRankAchievement!.rewardedPoints ?? 0
-            newAchievements[participant.id] = thirdRankAchievement!.id
-          }
-
-          return participant
-        })
-      }
-
-      // update xp of existing participants
-      promises = promises.concat(
-        existingParticipants
-          .filter(({ xp }) => typeof xp !== 'undefined')
-          .map(({ id, xp }) =>
-            ctx.prisma.participant.update({
-              where: { id },
-              data: {
-                xp: {
-                  increment: Number(xp),
-                },
-              },
-            })
-          )
-      )
-
-      // if the session is part of a course, update the course leaderboard with the accumulated points and award achievements
-      if (sessionLB && session.courseId) {
-        promises = promises.concat(
-          existingParticipants
-            .filter(
-              ({ score, hasParticipation }) =>
-                typeof score !== 'undefined' && hasParticipation
-            )
-            .map(({ id, score }) =>
-              ctx.prisma.leaderboardEntry.upsert({
-                where: {
-                  type_participantId_courseId: {
-                    type: 'COURSE',
-                    courseId: session.courseId!,
-                    participantId: id,
-                  },
-                },
-                include: {
-                  participation: true,
-                  participant: true,
-                },
-                create: {
-                  type: 'COURSE',
-                  course: {
-                    connect: {
-                      id: session.courseId!,
-                    },
-                  },
-                  participant: {
-                    connect: {
-                      id,
-                    },
-                  },
-                  participation: {
-                    connectOrCreate: {
-                      where: {
-                        courseId_participantId: {
-                          courseId: session.courseId!,
-                          participantId: id,
-                        },
-                      },
-                      create: {
-                        course: {
-                          connect: {
-                            id: session.courseId!,
-                          },
-                        },
-                        participant: {
-                          connect: {
-                            id,
-                          },
-                        },
-                      },
-                    },
-                  },
-                  score: score,
-                },
-                update: {
-                  score: {
-                    increment: score,
-                  },
-                },
-              })
-            )
-        )
-
-        // award new achievements
-        promises = promises.concat(
-          existingParticipants
-            .filter(({ id }) => typeof newAchievements[id] !== 'undefined')
-            .map(({ id }) =>
-              ctx.prisma.participant.update({
-                where: { id },
-                data: {
-                  achievements: {
-                    upsert: {
-                      where: {
-                        participantId_achievementId: {
-                          participantId: id,
-                          achievementId: newAchievements[id]!,
-                        },
-                      },
-                      create: {
-                        achievedAt: new Date(),
-                        achievedCount: 1,
-                        achievement: {
-                          connect: {
-                            id: newAchievements[id]!,
-                          },
-                        },
-                      },
-                      update: {
-                        achievedCount: {
-                          increment: 1,
-                        },
-                      },
-                    },
-                  },
-                },
-              })
-            )
-        )
-      }
-    }
-
-    // execute XP and points in the same transaction to prevent issues when one fails
-    // the session update later on should never fail, but we need the return value (keep separate)
-    await ctx.prisma.$transaction(promises)
-
-    const keys = await ctx.redisExec.keys(`s:${id}:*`)
-    const pipe = ctx.redisExec.multi()
-    for (const key of keys) {
-      pipe.unlink(key)
-    }
-    await pipe.exec()
-
-    const stoppedSession = await ctx.prisma.liveSession.update({
-      where: {
-        id,
-      },
-      data: {
-        status: SessionStatus.COMPLETED,
-        finishedAt: new Date(),
-        pinCode: null,
-      },
-    })
-
-    await sendTeamsNotifications(
-      'graphql/endSession',
-      `END Session ${session.name} with id ${session.id}.`
-    )
-
-    return stoppedSession
-  } catch (error) {
-    await sendTeamsNotifications(
-      'graphql/endSession',
-      `ERROR - failed to end session ${session.name} with id ${session.id}: ${error}`
-    )
-    throw error
-  }
-}
 
 interface GetCachedBlockResultsArgs {
   ctx: Context
